Create Fixxmlhttpoverload.ino
By adding the setTimeout function with a delay, you can control how often the XMLHttpRequest is sent to your ESP32. In this example, I've set the delay to 500 milliseconds, but you can adjust it to a value that suits your needs.

Repeat this modification for all the sliders to ensure a similar delay for each of them. This should help prevent the ESP32 from rebooting due to excessive requests.